requirements.txt
pandas
matplotlib
numpy
requests
requests-cache
tabulate
python-dotenv
Pillow
scikit-learn
cachetools
paho-mqtt
langchain
langchain-community
chromadb
sentence-transformers
pypdf
Flask
Flask-Cors
ollama
transformers
huggingface-hub
torch
huggingface-hub<1.0.0